# doc/OpenAPI 3.1.yaml
openapi: 3.1.0
info:
  title: Project Chimera - S12 Gateway API
  version: 0.3.0
servers:
  - url: https://chimera.example.com/api
security:
  - bearerAuth: []
paths:
...
  /experiments/run:
    post:
      summary: Run backtest/experiment
      operationId: experimentsRun
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExperimentsRunRequest" }
      responses:
        "202":
          description: Queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExperimentsRunResponse" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ExperimentsRunRequest:
      type: object
      properties:
        hypothesis_id: { type: string }
        design:
          type: object
          properties:
            window: { type: object, properties: { from: { type: string, format: date-time }, to: { type: string, format: date-time } }, required: [from, to] }
            metrics: { type: array, items: { type: string } }
      oneOf:
        - required: [hypothesis_id]
        - required: [design]
    ExperimentsRunResponse:
      type: object
      properties:
        exp_id: { type: string }
        status: { type: string, enum: [QUEUED, RUNNING, DONE, FAILED] }
...